Another way to join addresses to people.
A save saves 'child' records of the parent. It appears to be sensitive
to who was created when. In the first case, where I hung things from
the person, saving the person saved everything. But in this commit,
saving from the person didn't work, and I had to save the join model
AddressPerson.